+++ releasenotes.md
@@ -3,6 +3,7 @@
 ## Version 0.16.0
 ### New Features
 - IDNumber.Invalid()
+- IDNumber.SpanishCitizenNumber()
 - IDNumber.Valid()
 
 ## Version 0.15.0